Rename /benchmark to /read-write
