+++ read-write/AgentFactory.kt
@@ -14,8 +14,17 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.vaticle.typedb.benchmarks.storage.common
+package com.vaticle.typedb.benchmark.readwrite
 
-import com.vaticle.typedb.benchmark.framework.common.params.Config
+import com.vaticle.typedb.benchmark.readwrite.common.Context
+import com.vaticle.typedb.benchmark.framework.Agent
+import com.vaticle.typedb.benchmark.framework.typedb.TypeDBClient
 
-typealias Config = Config<ModelParams>
+class AgentFactory(client: TypeDBClient, context: Context): Agent.Factory() {
+    override val map: Map<Class<out Agent<*, *, *>>, () -> Agent<*, *, *>> = mapOf(
+        PersonAgent::class.java to { graph(client, context) },
+    )
+
+    fun graph(client: TypeDBClient, context: Context) = PersonAgent(client, context)
+
+}
