+++ read-write/Runner.kt
@@ -14,11 +14,11 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.vaticle.typedb.benchmarks.storage
+package com.vaticle.typedb.benchmark.readwrite
 
-import com.vaticle.typedb.benchmarks.storage.common.Config
-import com.vaticle.typedb.benchmarks.storage.common.Context
-import com.vaticle.typedb.benchmarks.storage.common.ModelParams
+import com.vaticle.typedb.benchmark.readwrite.common.Config
+import com.vaticle.typedb.benchmark.readwrite.common.Context
+import com.vaticle.typedb.benchmark.readwrite.common.ModelParams
 import com.vaticle.typedb.benchmark.framework.common.params.Database
 import com.vaticle.typedb.benchmark.framework.common.params.Options
 import com.vaticle.typedb.benchmark.framework.typedb.TypeDBClient
