+++ read-write/common/BUILD
