+++ read-write/common/Util.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.vaticle.typedb.benchmarks.storage.common
+package com.vaticle.typedb.benchmark.readwrite.common
 
 import com.vaticle.typedb.common.yaml.YAML
 import java.time.Duration
