+++ read-write/config/BUILD
@@ -17,7 +17,7 @@
 
 load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 
-package(default_visibility = ["//benchmark:__pkg__"])
+package(default_visibility = ["//read-write:__pkg__"])
 
 filegroup(
     name = "simulation_config",
